refactor: split project package service
